+++ src/main/resources/liquibase/db_changelog/create_table_balance.sql
@@ -2,6 +2,6 @@ CREATE TABLE Balance
 (
     id      BIGSERIAL PRIMARY KEY,
     user_id BIGINT NOT NULL,
-    coins   INTEGER,
+    coins   DECIMAL,
     FOREIGN KEY (user_id) REFERENCES Users (id)
 );
